### 2024-11-06 #7277 @Nikoleta-v3
Hello @evaaepelde,

Apologies for the delay in starting your submission; I’m currently in the process of moving to a new country. I’ll do my best to proceed as quickly as possible!

1. To get started, we need to find two reviewers. Could you please take a look at https://reviewers.joss.theoj.org/ and recommend up to four reviewers?

2. As you can see here https://github.com/openjournals/joss-reviews/issues/7277#issuecomment-2374520604, some references are missing DOIs. Could you please update your bibliography file? Once updated, you can run the command `@editorialbot check references` here to ensure everything is correct. If some references don’t have DOIs, please ignore the warning for now.

I had some time to go over the package, and I have a few comments:

### Contributing File

The repository is currently missing a CONTRIBUTING file, which I believe is essential for this project. If the long-term goal is to include data from countries beyond Spain, the CONTRIBUTING file should outline how others can contribute to expanding this dataset or adding other improvements. It should clearly explain how contributions can be made, whether through new analysis functions or data contributions, and specify that documentation should accompany any new functions.

To give you some ideas on structuring a CONTRIBUTING file, here’s an example from a recent publication I edited: [CONTRIBUTING.rst](https://forgemia.inra.fr/pherosensor/pherosensor-toolbox/-/blob/main/CONTRIBUTING.rst?ref_type=heads). It’s clear and to the point, so you may want to use it as a guide.

Other examples to consider:
- A more detailed example: [uqtestfuns](https://github.com/damar-wicaksono/uqtestfuns/blob/dev/CONTRIBUTING.MD)
- A very simple, one-sentence version: [stanscofi](https://github.com/RECeSS-EU-Project/stanscofi/)

While you can choose the style that best suits your project, I recommend including at least one section about contributing data.

Could you also ensure that the CONTRIBUTING file contains a section on how to run the tests?

### Data Included in Repository

Another issue is that the data is currently part of the repository. This isn’t typical, especially if the long-term goal of the package is to support data from multiple countries, as this could result in a large repository.

A common practice is to host the data elsewhere, with a function in the package to download it (accepting a path as an argument, for example). The user would only need to run this function the first time.

Is this something you’ve considered? Could you explain why the data is included in the repository?
