### 2024-12-09 #7277 @evaaepelde
Hi @Nikoleta-v3,

Thank you for the feedback and for taking the time to review our submission despite your move—congratulations and best of luck with settling in! Apologies for my delay in responding. I’ve been out of the office due to personal issues over the past month, but I’m now catching up and will address the points you’ve raised as quickly as possible.

1 - I’ve reviewed the reviewer suggestions page you shared and this is my proposal for the potential reviewers:
[Xin Zhao](https://reviewers.joss.theoj.org/reviewers/3426): [@realxinzhao](https://github.com/realxinzhao)
[Maria Gargiulo](https://reviewers.joss.theoj.org/reviewers/2262): [@thegargiulian](https://github.com/thegargiulian)
[@chrisaberson](https://reviewers.joss.theoj.org/reviewers/2019): [@chrisaberson](https://github.com/chrisaberson)
[Blake Rayfield](https://reviewers.joss.theoj.org/reviewers/2222): [@bkrayfield](https://github.com/bkrayfield)
[Erik-Jan van Kesteren](https://reviewers.joss.theoj.org/reviewers/3295): [@vankesteren](https://github.com/vankesteren)

2 - Regarding the references, I’ll update the bibliography file to include missing DOIs and run the @editorialbot check references command to verify today.

3 - We find the idea of including a CONTRIBUTING file very compelling. However, our current plan is to focus on updating the package over the coming months to include data from all EU27 countries. We aim to incorporate the CONTRIBUTING file as part of this second version, ensuring it aligns with the expanded scope and provides clear guidance for future contributions.

4 - About the data included in the repository: Initially, we attempted to follow the approach you suggested: hosting the data in an external repository and creating a function within the package to download it. However, the data size proved to be too large, causing the function to fail as it took too long to download. While we eventually fixed the issue, the process still remained slow and impractical for users, so we decided to include the data directly in the package to avoid unnecessary delays and improve the user experience. That said, as part of our second version, where we plan to extend the dataset to include all EU27 countries, we aim to revisit and improve this aspect. Our goal is to implement a more efficient solution that balances usability and scalability.

Thank you again for the thorough review.

Best regards,

Eva
