### 2024-12-10 48b83ea
add missing DOIs
