### 2024-12-10 fec39aa
add missing ISBNs
