+++ paper/paper.bib
@@ -52,6 +52,7 @@
 	publisher = {Vintage},
 	author = {Davis, A.Y.},
 	year = {1983},
+	isbn = {978-0-307-79849-7}
 }
 
 @article{djoudi2016,
@@ -124,7 +125,8 @@
 @book{bazoli2022,
   title={Study on distributional impact assessment},
   author={Bazoli, Nicola and Burlacu, Sergiu and Fiorio, Carlo and Marzadro, Sonia and Pycroft, Jonathan and Vergolini, Loris and others},
-  year={2022}
+  year={2022},
+  doi = {https://data.europa.eu/doi/10.2767/511644}
 }
 
 @article{walker2010,
@@ -164,6 +166,7 @@
 	author = {Eurostat},
 	year = {2003},
 	file = {HOUSEHOLD BUDGET SURVEYS IN THE EU — Methodology a.pdf:C\:\\Users\\eva.alonso\\Zotero\\storage\\8AZP2MM5\\HOUSEHOLD BUDGET SURVEYS IN THE EU — Methodology a.pdf:application/pdf},
+	isbn = {92-894-5435-0}
 }
 
 @article{cazcarro2022,
