+++ pkgdown.yml
@@ -10,4 +10,4 @@ articles:
   Tutorials: Tutorials.html
   TutorialsEnergyPoverty: TutorialsEnergyPoverty.html
   TutorialsTransportPoverty: TutorialsTransportPoverty.html
-last_built: 2024-12-10T09:44Z
+last_built: 2024-12-10T12:52Z
