### 2024-12-10 0153aaa
fix invalid DOIs
