+++ paper/paper.bib
@@ -115,7 +115,7 @@
   title={Anticipating futures through models: the rise of Integrated Assessment Modelling in the climate science-policy interface since 1970},
   author={Van Beek, Lisette and Hajer, Maarten and Pelzer, Peter and van Vuuren, Detlef and Cassen, Christophe},
   journal={Global Environmental Change},
-  doi = {https://doi.org/10.1016/j.gloenvcha.2020.102191},
+  doi = {10.1016/j.gloenvcha.2020.102191},
   volume={65},
   pages={102191},
   year={2020},
@@ -126,14 +126,14 @@
   title={Study on distributional impact assessment},
   author={Bazoli, Nicola and Burlacu, Sergiu and Fiorio, Carlo and Marzadro, Sonia and Pycroft, Jonathan and Vergolini, Loris and others},
   year={2022},
-  doi = {https://data.europa.eu/doi/10.2767/511644}
+  doi = {10.2767/511644}
 }
 
 @article{walker2010,
   title={Environmental justice, impact assessment and the politics of knowledge: The implications of assessing the social distribution of environmental outcomes},
   author={Walker, Gordon},
   journal={Environmental impact assessment review},
-  doi = {https://doi.org/10.1016/j.eiar.2010.04.005},
+  doi = {10.1016/j.eiar.2010.04.005},
   volume={30},
   number={5},
   pages={312--318},
@@ -173,7 +173,7 @@
   title={Linking multisectoral economic models and consumption surveys for the European Union},
   author={Cazcarro, Ignacio and Amores, Antonio F and Arto, Inaki and Kratena, Kurt},
   journal={Economic Systems Research},
-  doi = {https://doi.org/10.1080/09535314.2020.1856044},
+  doi = {10.1080/09535314.2020.1856044},
   volume={34},
   number={1},
   pages={22--40},
@@ -185,7 +185,7 @@
   title={Transport poverty indicators: A new framework based on the household budget survey},
   author={Alonso-Epelde, E and Garc{\'\i}a-Muros, X and Gonz{\'a}lez-Eguino, M},
   journal={Energy Policy},
-  doi = {https://doi.org/10.1016/j.enpol.2023.113692},
+  doi = {10.1016/j.enpol.2023.113692},
   volume={181},
   pages={113692},
   year={2023},
