### 2024-12-11 #7277 @evaaepelde
Hi @Nikoleta-v3,

I wanted to update you on the bibliography. I have added the missing DOIs where available, and for the three references that do not have a DOI, I’ve included the ISBN instead. Please let me know if there’s anything else you need.

Thank you for your guidance!

Best regards,

Eva
