### 2025-01-04 #7277 @Nikoleta-v3
Hello @evaaepelde 👋🏻

Apologies for the delay in my response. I’ve now fully moved, so hopefully, there won’t be any more hiccups on my end.

Regarding the mission DOI, the book does have one: https://doi.org/10.4324/9780203060902. Please include that, since the chapter itself doesn’t have a DOI.

Thank you for addressing my comments. It sounds like version two will be an even bigger and more complete package. This naturally raises the question: why not wait and publish that version instead?

I’m not sure, maybe these are just my personal views. So, I’m going to go ahead and invite some people to review the paper to see if these concerns come up again. Thank you for suggesting potential reviewers 😄
