### 2025-01-04 #7277 @thegargiulian
Happy new year, @Nikoleta-v3!

I'll be available in the coming weeks to review this paper.
