### 2025-01-13 #7277 @jonsampedro
Hi @Nikoleta-v3, happy new year!

I work closely with the creator of this tool, so I believe I can act as a reviewer this time, considering the potential conflict of interest...

That said, I’d be happy to participate as a reviewer on a different occasion.

Best regards,
Jon
