### 2025-01-13 #7277 @kanishkan91
@Nikoleta-v3 I'm currently editing a few articles for JOSS. I can provide a review of this package, but it might take me 4 weeks instead of 2. If that's ok, feel free to assign me.
