### 2025-01-16 #7277 @MatthewHeun
Thanks for the offer.  However, I have no expertise in the purpose of the package, equity and distribUtional impacts for socioeconomic analysis.  I think I should also decline.  Thank you, though.

Cheers,

Matt

> On Jan 16, 2025, at 5:17 AM, Hanne Oberman ***@***.***> wrote:
>
>
> 👋🏻 @realxinzhao <https://github.com/realxinzhao> @thegargiulian <https://github.com/thegargiulian> @chrisaberson <https://github.com/chrisaberson> @hanneoberman <https://github.com/hanneoberman> happy new year! Would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission I'd like you to review is titled: "MEDUSA: An R tool to model equity and distributional impacts". You can find more information at the top of this Github issue 🆙
>
> Please let me know if you're available 😄 Thank you!
>
> Hi! Apologies for the delayed response. I don't feel confident reviewing this specific topic, so I’ll have to pass. Thanks for understanding!
>
