### 2025-01-17 #7277 @Nikoleta-v3
👋🏻 @sdmccabe @apwheele @cjbarrie happy new year! Would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "MEDUSA: An R tool to model equity and distributional impacts". You can find more information at the top of this Github issue 🆙
