### 2025-01-20 #7697 @editorialbot
[REVIEW]: MEDUSA: An R tool to model equity and distributional impacts
**Submitting author:** @evaaepelde (<a href="http://orcid.org/0000-0002-4052-1703">Eva Alonso-Epelde</a>)
**Repository:** https://github.com/bc3LC/medusa
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1.0
