### 2025-01-20 #7697 @Nikoleta-v3
Hey all this is the review thread for the paper. All of our communications will happen here from now on.

For the reviewers, the first step is to create a checklist for your review by entering

These checklists contain the JOSS requirements ✅ As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.

Please feel free to ping me (@Nikoleta-v3) if you have any questions/concerns. 😄 🙋🏻
