### 2025-02-09 #7697 @apwheele
So the flow to work with the software + the vignette I find a bit strange. The example wants you to download a csv file, edit it, and then use the dataframe. I would have an option to just return the dataframe directly. But probably more important is to show in the vignette an example of the difference in outputs if say you change one of the coicop variables shocks.

Minor, install should IMO not state RStudio as a dependency. I use conda, and on windows from an environment `conda create --name RMedusa -c conda-forge r-base>=4.1 r-remotes rtools`, then I can do from the R terminal `library(remotes); install_github("bc3LC/medusa")` works for me (I just presume at this point I need Rtools on windows for everything)

Also no community guidelines and state of the field (state of the field may simply be there is nothing comparable)
