### 2025-02-24 #7697 @Nikoleta-v3
@evaaepelde Congratulations on your defense! Should we be referring to you as Doctor now? 😄🥳

Thank you. I just wanted to make sure you were keeping an eye on this.
