### 2025-03-03 f49e05c
add community guidelines and state of the field
