### 2025-03-03 4c5e08e
add community guidelines and state of the field
