+++ pkgdown.yml
@@ -10,4 +10,4 @@ articles:
   Tutorials: Tutorials.html
   TutorialsEnergyPoverty: TutorialsEnergyPoverty.html
   TutorialsTransportPoverty: TutorialsTransportPoverty.html
-last_built: 2025-03-03T17:16Z
+last_built: 2025-03-03T17:20Z
