+++ paper/paper.bib
@@ -202,3 +202,11 @@
 	year = {2016},
 	file = {Rivisteweb\: Article Details:C\:\\Users\\eva.alonso\\Zotero\\storage\\AJLBX8RM\\84373.html:text/html},
 }
+
+@article{weyant2017,
+  title={Some contributions of integrated assessment models of global climate change},
+  author={Weyant, John},
+  journal={Review of Environmental Economics and Policy},
+  year={2017},
+  publisher={The University of Chicago Press}
+}
