### 2025-03-05 210d1c4
add data links and modify Package data vignette
