+++ paper/paper.md
@@ -60,7 +60,7 @@ Addressing critical challenges like climate change requires ambitious policies t
 
 `medusa` facilitates distributional impact analyses through an overnight-effect microsimulation model, leveraging microdata from the Household Budget Survey (HBS) [^1], a standardized and comprehensive dataset available across EU countries [@eurostat2003]. The HBS offers detailed insights into household consumption patterns and socioeconomic characteristics at both household and individual levels, allowing for highly granular analysis. This enables the integration of an intersectional approach[^2] considering factors such as class, gender, and race, and provides more robust and nuanced results for assessing policy impacts on diverse population groups.
 
-[^1]: Since the model currently operates for Spain, the microdata used in this package come from the Spanish National Statistics Institute (INE). They can be accessed at [this link](https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736176806&menu=resultados&idp=1254735976608#_tabs-1254736195147).
+[^1]: Since the package currently operates for Spain, the microdata used in this package come from the Spanish National Statistics Institute (INE). They can be accessed at [this link](https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736176806&menu=resultados&idp=1254735976608#_tabs-1254736195147).
 
 [^2]: Intersectionality refers to the fact that the privileges or oppression of each individual depend on the multiple social categories to which he or she belongs, which are social constructs and can change over time [@cho2013; @crenshaw1994; @davis1983; @djoudi2016; @kaijser2014]. Intersectionality is therefore also a tool for analysing the articulation of different socio-economic categories (e.g. class, gender, race, etc.) rather than considering them as independent forms of power relations [@colombo2016].
 
