+++ vignettes/TutorialsExample.Rmd
@@ -1,0 +1,134 @@
+---
+title: "Step-by-step example"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{TutorialsExample}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+# Step by step example
+
+## Introduction
+
+In this example, we analyze the distributional impacts of energy price shocks in Spain, focusing on two scenarios:
+
+- **Scenario 1**: A price shock affecting fuel prices for private transportation, similar to what occurred during the 2022 energy crisis.
+- **Scenario 2**: A policy intervention where the government implements a measure to reduce domestic energy prices to mitigate the crisis.
+
+We will use the `medusa` package to evaluate the socioeconomic impacts of these scenarios.
+
+## Step 1: Calculate Distributional Impacts by Income Level
+
+### Define the Scenarios
+
+1. Download the example file into your working directory by running:
+
+```r
+ex_shocks()
+```
+
+2. Locate and open the CSV file **"Example_shocks.csv"** in your working directory.
+
+<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/ex_shocks.PNG" title="Example shocks file" alt="ex_shocks" width="80%" height="80%"/>
+
+3. Each column labeled `Scenario__` represents a different scenario. We will keep Scenario 1 and Scenario 2 (but you can also rename them to be more intuitive).
+
+4. Enter the price shocks for each **COICOP** category and each Scenario. In this case, we introduce:
+   - `1.5` in **"Use of personal vehicles"** for both scenarios to indicate that fuel prices increase by 50%.
+   - `0.5` in **"Electricity, gas and other fuels"** in **Scenario 2** to indicate that, due to the government policy, domestic energy prices decrease by 50%.
+   - Mantain `1` in all other categories to indicate that prices in these categories remain unaffected.
+
+<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/ex_shocks2.PNG" title="Example shocks csv" alt="ex_shocks2" width="80%" height="80%"/>
+
+5. Save the edited CSV file.
+
+6. Upload the edited file into R:
+
+```r
+exampledf <- read.csv("Example_shocks.csv", header = TRUE, sep = ",", dec = ".")
+```
+
+### Run Distributional Impact Calculation
+
+7. Identify the variables available for distributional impact analysis:
+
+```r
+available_var_impact()
+```
+
+This function will return the set of available variables in the console. For this example, we will select **"DECILE"**, which refers to household income deciles. For more information on available variables, visit [here](https://bc3lc.github.io/medusa/articles/AvailableVariables.html).
+
+<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/available_var.PNG" title="Available variables" alt="available_var" width="80%" height="80%"/>
+
+8. Compute the distributional impacts by income decile for 2019 data:
+
+```r
+calc_di(2019,                      # Base year: 2019
+        shocks = exampledf,        # File containing the shocks
+        var_impact = "DECILE")     # Select income deciles as the socioeconomic variable
+```
+
+<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/di_decile.PNG" title="Distributional impats: R outputs" alt="di_decile" width="80%" height="80%"/>
+
+By default, this function saves the produced results and the corresponding figure in your working directory for each specified variable, in this case, for **DECILE**.
+
+<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/di_decile2.PNG" title="Distributional impats: outputs in the wd" alt="di_decile2" width="80%" height="80%"/>
+
+### Results Interpretation
+
[…]
