+++ pkgdown.yml
@@ -9,5 +9,6 @@ articles:
   TransportPoverty: TransportPoverty.html
   Tutorials: Tutorials.html
   TutorialsEnergyPoverty: TutorialsEnergyPoverty.html
+  TutorialsExample: TutorialsExample.html
   TutorialsTransportPoverty: TutorialsTransportPoverty.html
-last_built: 2025-03-07T10:28Z
+last_built: 2025-03-07T11:15Z
