### 2025-03-07 2c983a8
Revert "try fix figures. png extention"
This reverts commit 18ed6cb1c893f35b0a4ec05d237548564b6fbaa3.
