+++ vignettes/TutorialsExample.Rmd
@@ -1,19 +1,19 @@
 ---
-  title: "Step-by-step example"
+title: "Step-by-step example"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{TutorialsExample}
-%\VignetteEngine{knitr::rmarkdown}
-%\VignetteEncoding{UTF-8}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
 ---
-  
-  # Step by step example
-  
-  ## Introduction
-  
-  In this example, we analyze the distributional impacts of energy price shocks in Spain, focusing on two scenarios:
-  
-  - **Scenario 1**: A price shock affecting fuel prices for private transportation, similar to what occurred during the 2022 energy crisis.
+
+# Step by step example
+
+## Introduction
+
+In this example, we analyze the distributional impacts of energy price shocks in Spain, focusing on two scenarios:
+
+- **Scenario 1**: A price shock affecting fuel prices for private transportation, similar to what occurred during the 2022 energy crisis.
 - **Scenario 2**: A policy intervention where the government implements a measure to reduce domestic energy prices to mitigate the crisis.
 
 We will use the `medusa` package to evaluate the socioeconomic impacts of these scenarios.
@@ -23,8 +23,8 @@ We will use the `medusa` package to evaluate the socioeconomic impacts of these 
 ### Define the Scenarios
 
 1. Download the example file into your working directory by running:
-  
-  ```r
+
+```r
 ex_shocks()
 ```
 
@@ -32,90 +32,86 @@ ex_shocks()
 
 <img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/ex_shocks.PNG" title="Example shocks file" alt="ex_shocks" width="80%" height="80%"/>
 
-adsf
-
-<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/ex_shocks22.png" title="Example shocks file" alt="ex_shocks" width="80%" height="80%"/>
-  
-  3. Each column labeled `Scenario__` represents a different scenario. We will keep Scenario 1 and Scenario 2 (but you can also rename them to be more intuitive).
+3. Each column labeled `Scenario__` represents a different scenario. We will keep Scenario 1 and Scenario 2 (but you can also rename them to be more intuitive).
 
 4. Enter the price shocks for each **COICOP** category and each Scenario. In this case, we introduce:
-  - `1.5` in **"Use of personal vehicles"** for both scenarios to indicate that fuel prices increase by 50%.
-- `0.5` in **"Electricity, gas and other fuels"** in **Scenario 2** to indicate that, due to the government policy, domestic energy prices decrease by 50%.
-- Mantain `1` in all other categories to indicate that prices in these categories remain unaffected.
+   - `1.5` in **"Use of personal vehicles"** for both scenarios to indicate that fuel prices increase by 50%.
+   - `0.5` in **"Electricity, gas and other fuels"** in **Scenario 2** to indicate that, due to the government policy, domestic energy prices decrease by 50%.
+   - Mantain `1` in all other categories to indicate that prices in these categories remain unaffected.
 
 <img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/ex_shocks2.PNG" title="Example shocks csv" alt="ex_shocks2" width="80%" height="80%"/>
-  
-  5. Save the edited CSV file.
+
+5. Save the edited CSV file.
 
 6. Upload the edited file into R:
-  
-  ```r
+
+```r
 exampledf <- read.csv("Example_shocks.csv", header = TRUE, sep = ",", dec = ".")
 ```
 
 ### Run Distributional Impact Calculation
 
 7. Identify the variables available for distributional impact analysis:
-  
-  ```r
+
+```r
[…]
