+++ pkgdown.yml
@@ -11,4 +11,4 @@ articles:
   TutorialsEnergyPoverty: TutorialsEnergyPoverty.html
   TutorialsExample: TutorialsExample.html
   TutorialsTransportPoverty: TutorialsTransportPoverty.html
-last_built: 2025-03-07T12:09Z
+last_built: 2025-03-07T12:26Z
