+++ vignettes/TutorialsExample.Rmd
@@ -30,7 +30,7 @@ ex_shocks()
 
 2. Locate and open the CSV file **"Example_shocks.csv"** in your working directory.
 
-<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/ex_shocks.PNG" title="Example shocks file" alt="ex_shocks" width="80%" height="80%"/>
+<img src="https://raw.githubusercontent.com/bc3LC/medusa/main/vignettes/figs/ex_shocks.png" title="Example shocks file" alt="ex_shocks" width="80%" height="80%"/>
 
 3. Each column labeled `Scenario__` represents a different scenario. We will keep Scenario 1 and Scenario 2 (but you can also rename them to be more intuitive).
 
@@ -39,7 +39,7 @@ ex_shocks()
    - `0.5` in **"Electricity, gas and other fuels"** in **Scenario 2** to indicate that, due to the government policy, domestic energy prices decrease by 50%.
    - Mantain `1` in all other categories to indicate that prices in these categories remain unaffected.
 
-<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/ex_shocks2.PNG" title="Example shocks csv" alt="ex_shocks2" width="80%" height="80%"/>
+<img src="https://raw.githubusercontent.com/bc3LC/medusa/main/vignettes/figs/ex_shocks2.png" title="Example shocks csv" alt="ex_shocks2" width="80%" height="80%"/>
 
 5. Save the edited CSV file.
 
@@ -59,7 +59,7 @@ available_var_impact()
 
 This function will return the set of available variables in the console. For this example, we will select **"DECILE"**, which refers to household income deciles. For more information on available variables, visit [here](https://bc3lc.github.io/medusa/articles/AvailableVariables.html).
 
-<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/available_var.PNG" title="Available variables" alt="available_var" width="80%" height="80%"/>
+<img src="https://raw.githubusercontent.com/bc3LC/medusa/main/vignettes/figs/available_var.png" title="Available variables" alt="available_var" width="80%" height="80%"/>
 
 8. Compute the distributional impacts by income decile for 2019 data:
 
@@ -69,11 +69,11 @@ calc_di(2019,                      # Base year: 2019
         var_impact = "DECILE")     # Select income deciles as the socioeconomic variable
 ```
 
-<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/di_decile.PNG" title="Distributional impats: R outputs" alt="di_decile" width="80%" height="80%"/>
+<img src="https://raw.githubusercontent.com/bc3LC/medusa/main/vignettes/figs/di_decile.png" title="Distributional impats: R outputs" alt="di_decile" width="80%" height="80%"/>
 
 By default, this function saves the produced results and the corresponding figure in your working directory for each specified variable, in this case, for **DECILE**.
 
-<img src="https://github.com/bc3LC/medusa/blob/main/vignettes/figs/di_decile2.PNG" title="Distributional impats: outputs in the wd" alt="di_decile2" width="80%" height="80%"/>
+<img src="https://raw.githubusercontent.com/bc3LC/medusa/main/vignettes/figs/di_decile2.png" title="Distributional impats: outputs in the wd" alt="di_decile2" width="80%" height="80%"/>
 
 ### Results Interpretation
 
@@ -82,7 +82,7 @@ The figure generated by `calc_di` illustrates the relative impact (%) on total e
 - In **Scenario 1** (left panel), all deciles experience a significant welfare loss, but the impact is higher on **middle- and high-income households**. This suggests that these households, which rely more on private transportation, are more exposed to fuel price increases.
[…]
