### 2025-03-12 a30c277
add check vars + fix readme and tutorials
