+++ R/main.R
@@ -61,6 +61,9 @@ calc_di <-function(year, elevate=F, shocks, var_impact = "all", var_intersec = N
   # Check year parameter
   check_year(year)
 
+  # Check var_impact
+  check_var_impact(var_impact)
+
   # get hbs files
   hbs <- get(paste0("epf_list_", year))
 
@@ -105,7 +108,7 @@ calc_di <-function(year, elevate=F, shocks, var_impact = "all", var_intersec = N
   # calculate distributional impacts
   if(!is.null(var_impact)){
 
-    if(var_impact == "all"){
+    if(any(var_impact == "all")){
       var_impact = categories$categories
       }
     di <- impact(epf, var = var_impact, save = save,
