### 2025-03-12 #7697 @evaaepelde
[thegargiulian](https://github.com/thegargiulian) and [@apwheele](https://github.com/apwheele), thank you very much for your detailed comments and suggestions. Your review has been extremely helpful in improving both the manuscript and the package presented. Below, I address each of your comments point by point:

**1. So the flow to work with the software + the vignette I find a bit strange. The example wants you to download a csv file, edit it, and then use the dataframe. I would have an option to just return the dataframe directly. But probably more important is to show in the vignette an example of the difference in outputs if say you change one of the coicop variables shocks. ([@apwheele](https://github.com/apwheele))**
Thank you for your valuable feedback. Following your comment and the suggestion from [thegargiulian](https://github.com/thegargiulian), we have added a new section that walks users through an step-by-step example, explaining the outputs at each stage. We believe this will be very useful for users, so we really appreciate your input. You can find this new section under Tutorials > [Step-by-step example](https://bc3lc.github.io/medusa/articles/TutorialsExample.html).
Regarding the need to download and edit a CSV file, we understand that this approach may not be the most common. However, the package is designed to be accessible to a wide range of users, including those with little experience in R. Given the large number of COICOP categories and the potential unfamiliarity with coding, we believe this method provides the most straightforward way to introduce price shocks for multiple scenarios. That said, we acknowledge the value of offering a more direct alternative for advanced users. As we begin working on the next version of the package this month, we will explore options to implement this functionality in a way that does not require downloading and uploading CSV files.

**2. Minor, install should IMO not state RStudio as a dependency. I use conda, and on windows from an environment conda create --name RMedusa -c conda-forge r-base>=4.1 r-remotes rtools, then I can do from the R terminal library(remotes); install_github("bc3LC/medusa") works for me (I just presume at this point I need Rtools on windows for everything). ([@apwheele](https://github.com/apwheele))**
Thank you for your comment. We agree that RStudio is not a strict dependency for installing the package. In the installation guide, our aim was to provide a user-friendly approach for those who may not be familiar with R, which is why we included RStudio as a reference in the first step. However, as you correctly point out, more advanced users can install the package using alternative methods, such as the one you describe. To ensure clarity, we have revised the installation instructions to explicitly mention that RStudio is not required and that other installation methods are possible.

```{r}
To use `medusa` package first you need to follow this steps:

1.  Install R (required) and optionally an interface to facilitate usage, such as RStudio:

    -   To download R click [here](https://www.r-project.org/)

    -   To download Rstudio click [here](https://www.rstudio.com/)

```

**3. Also no community guidelines and state of the field (state of the field may simply be there is nothing comparable). ([@apwheele](https://github.com/apwheele))**
Community Guidelines section and a State of the Field section in the [README](https://bc3lc.github.io/medusa/index.html). We appreciate your feedback, as these additions help improve the clarity and usability of the package.

**4. General checks: Can you include a link in the paper to the actual source of the EuroStat data, in addition to the 2003 reference? ([thegargiulian](https://github.com/thegargiulian))**
 I have now added a footnote with a link to the microdata source from the Spanish National Statistics Institute (INE), as the current version of the model is based on data for Spain. For the next version of the package, we plan to use Eurostat microdata and will detail the corresponding link.

**5. Functionality: In working through the ex_shocks() part of the vignette for calculating distributional impacts, I get the following warning: Warning message: In impact(epf, var = var_impact, save = save, file_name = file_name_impact,  : WORKDAYRP is not present in the dataset ([thegargiulian](https://github.com/thegargiulian))**
Thank you so much for spotting this warning. It appears because in the survey year for which you are performing the calculations, the variable WORKDAYRP was not collected. The message simply informs you that results for this variable will not be available. This is not expected to be a common issue, as most users will likely work with the most recent surveys to analyze distributional impacts, and in recent years, all relevant variables have been included. Additionally, only a couple of variables are missing in the earliest surveys of the time series. Let me know if you encounter any other issues!

**6. Functionality: This example gives me an error (my test)**
```{r}
exampledf <- read.csv("Example_shocks.csv", header = TRUE, sep = ",", dec = ".")
#I did not add any additional scenarios:
vars <- c("ZONA", "TAMAMU")
#sub in 2010 for year
calc_di(2010, elevate = FALSE, shocks = exampledf, var_impact = vars)
Error in if (var_impact == "all") { : the condition has length > 1 ([thegargiulian](https://github.com/thegargiulian))
```
Thank you for catching this! The error is because the variables TAMAMU and ZONA are not existing, as they must be provided in English. The issue arose because, by mistake, I used the Spanish versions of these variable names in the vignettes, but I've now corrected it. I really appreciate your feedback! Additionally, I've updated the calc_di() function to include a new error message when a user inputs a variable that doesn't exist, making it clearer and avoiding future errors. Let me know if you run into any further issues!

**7. Functionality:  ex_var_intersec() does not return anything ([thegargiulian](https://github.com/thegargiulian))**
The function ex_var_intersec() does not return an output in R; instead, it saves a CSV file in your working directory. If you want to see the available variables directly in R, you should use available_var_intersec(). We acknowladge that it was not very intuitive; thus, we have added a message that will appear when the function ex_var_intersec() runs indicating where the CSV file has been saved.

**8. Functionality: I agree with [@apwheele](https://github.com/apwheele) that the workflow of editing csvs isn't super intuitive ([thegargiulian](https://github.com/thegargiulian))**
Thank you for your feedback. We understand that downloading and editing a CSV file may not be the most intuitive approach. However, MEDUSA is designed to be accessible to a wide range of users, including those with little experience in R. Given the large number of COICOP categories and the potential unfamiliarity with coding, we believe this method provides the most straightforward way to introduce price shocks for multiple scenarios. That said, we recognize the importance of offering alternative workflows for more advanced users. As we begin working on the next version of the model this month, we will explore ways to make this process more flexible, reducing the need to manually edit CSV files while maintaining ease of use.

**9. Functionality: In general, I think the vignettes would be more effective for users wishing to use the package if you could demonstrate how the package can be used to answer a particular research question and then work through using all of the functions to answer that question. ([thegargiulian](https://github.com/thegargiulian))**
Based on your feedback and that of [@apwheele](https://github.com/apwheele), we have added a [Step-by-step example](https://bc3lc.github.io/medusa/articles/TutorialsExample.html) section in the Tutorials, where we guide users through a complete workflow, demonstrating how MEDUSA can be used to answer a specific research question. This example walks through each step of the analysis, also explaining the outputs. We believe this addition will greatly enhance the usability of the package, making it easier for users to understand how to apply MEDUSA in their own research/analysis. We appreciate your insightful comment!

**10. Documentation: Can you add a bit of information about the state of the field to the repository? ([thegargiulian](https://github.com/thegargiulian))**
Thank you for your suggestion. We have now included a State of the field section in the [README](https://bc3lc.github.io/medusa/index.html), providing context on the importance of distributional analysis in policy assessments and highlighting how MEDUSA fills a critical gap as the first R package of its kind. This addition helps clarify the motivation behind the package and its relevance in the field.

**11. Documentation:  The repository is currently missing information on how individuals can contribute to the package, report issues, and seek support ([thegargiulian](https://github.com/thegargiulian))**
We have now added a Contributing Guidelines section to the [README](https://bc3lc.github.io/medusa/index.html), outlining how users can report issues, provide feedback, and contribute to the development of MEDUSA. This section includes information on opening issues, submitting pull requests, and seeking support. We appreciate your input in making the repository more user-friendly!

**12. Documentation: Functions look to be sufficiently documented, but I think the data included in the package needs a bit more transparency and documentation. Data objects should include a link to the original source (e.g., data from HBS should have a link to the HBS even if you're providing a cleaned version). I also think it would be helpful to follow the approach outlined in [_R Packages (2e)_](https://r-pkgs.org/data.html) in keeping the original data/scripts in data-raw folder to preserve all of the cleaning operations you undertook (see especially Section 7.1.1). I think that this is especially important because I don't see the cleaning process documented elsewhere in the repository or the paper. ([thegargiulian](https://github.com/thegargiulian))**
Thank you for this thoughtful comment. The adjustments made to the data are documented in the vignette ["Package Data"](https://bc3lc.github.io/medusa/articles/PackageData.html), where we describe the modifications applied before including the dataset in the package. These adjustments are minimal ---primarily merging the expenditure and household files and adding a few extra variables, which we explicitly mention in that section. Any further data adjustments, such as aligning expenditures with national accounts, are only performed when users explicitly choose to do so via the calc_di() function. These transformations are handled internally using the load_rawhbs() function, which is included in the package.
Initially, we intended to keep the original raw data within the package, but given the large volume of information, running the function was extremely slow, making it impractical for users. For this reason, we decided to pre-process the data externally and include the modified version in the package instead. However, users who wish to work with the raw microdata from INE (linked in the paper) can do so by running the load_rawhbs() function themselves.
To improve clarity, I have revised the vignette section, and I hope the process is now more transparent. In any case, this data section will soon undergo further changes, as the second version of the package will use Eurostat data, requiring additional adjustments to harmonize surveys across the EU-27 countries.

**13. Software paper:  In the Summary, maybe add a relevant citation for IAMs, a non-specialist might not be familiar with what these are and it would be nice to have a reference to look at ([thegargiulian](https://github.com/thegargiulian))**
Thank you for the suggestion! I have now included a citation to this [article](https://ideas.repec.org/a/oup/renvpo/v11y2017i1p115-137..html), which I believe provides a useful introduction to Integrated Assessment Models (IAMs) for non-specialists.

**14. Software paper:   Any existing software to do this type of analysis? Knowing a little bit more about existing software would help me understand the contribution better (I have not doubt that there's a contribution here, I just want to understand more about its nature!).([thegargiulian](https://github.com/thegargiulian))**
To the best of our knowledge, there was no existing software specifically designed for conducting distributional impact assessments of policies in an accessible way. While some economic models allow for similar analyses, they are often highly specialized, require advanced technical expertise, or are not openly available. MEDUSA fills this gap by providing the first R package dedicated to this type of analysis, making it more accessible to researchers, policymakers, and analysts. By offering a user-friendly framework, MEDUSA enables a broader audience to incorporate equity considerations into policy evaluation, which has been a missing component in many existing approaches. We appreciate your interest in understanding the contribution of the package!

**15. Software paper:  I would also recommend [_Data Feminism_](https://data-feminism.mitpress.mit.edu/) as a reference for intersectional quantitative work (not necessarily to cite, but maybe for personal and scholarly interest) ([thegargiulian](https://github.com/thegargiulian))**
Thank you for this recommendation! One of my main research objectives is to integrate a gender perspective from an intersectional approach into socioeconomic analyses, so this reference will be extremely useful both for my current work and future projects. I really appreciate the suggestion!

I truly appreciate your time and effort in reviewing this work. Please let me know if there are any further comments or clarifications needed.

Best regards,

Eva Alonso-Epelde
