### 2025-03-13 ed7ef31
improve check vars function + update tutorials
