+++ vignettes/Tutorials.Rmd
@@ -52,7 +52,7 @@ calc_di( year,                     # Base year for the simulation (available 200
 ```r
 available_var_impact()
 ```
-2. Select one of the variables that the function returns in the console. For example, "ZONE".
+2. Select one of the variables that the function returns in the console. For example, "DECILE".
 
 3. Enter that variable in the `var_impact` argument of `calc_di`. By default, `var_impact = "all"` , i.e. it calculates the distributional impacts for all variables returned by `available_var_impact()`. For more information on variables click [here](https://bc3lc.github.io/medusa/articles/AvailableVariables.html). 
  ``` r
