### 2025-03-14 e8cb310
fix check var_impact
