### 2025-03-17 1342835
fix check var_impact
