+++ R/main.R
@@ -61,7 +61,7 @@ calc_di <-function(year, elevate=F, shocks, var_impact = "all", var_intersec = N
   # Check year parameter
   check_year(year)
 
-  if (var_impact != "all") {
+  if (!is.null(var_impact) && var_impact != "all") {
     # Check var_impact
     check_var_impact(var_impact)
   }
