### 2025-03-20 4d7b6b8
start rawhbs_eu + check year and country
