+++ R/main_eu.R
@@ -1,0 +1,13 @@
+#' country_code
+#'
+#' Function that returns the country codes for EU27
+#' @return available countries of the package and their codes
+#' @export
+country_code <- function(){
+  av_country <- country
+  print(av_country)
+}
+
+
+
+
