### 2025-03-20 d1a520f
rawhbs_eu: select country codes of the available data
