### 2025-03-20 fa34b5c
rawhbs_eu: update paths to read data
