+++ DESCRIPTION
@@ -30,6 +30,7 @@ Imports:
   	tidyr,
   	ggplot2,
   	stringr,
+	openxlsx,
   	here,
 	  spatstat.geom,
 	  tibble
