### 2025-03-20 785ca5c
improve rawhbs_eu
