+++ inst/extdata/saveDataFiles.R
@@ -171,5 +171,5 @@ use_data(graph_labels, overwrite=T)
 rawDataFolder = paste0(here::here(),"/inst/extdata/eu/")
 
 # country
-country = read.csv(paste0(rawDataFolder,"country.csv"),header=T)
-use_data(country, overwrite=T)
+country = read.csv(paste0(rawDataFolder,"av_country.csv"),header=T)
+use_data(av_country, overwrite=T)
