+++ R/utils_eu.R
@@ -2,6 +2,59 @@ library(usethis)
 library(magrittr)
 options(dplyr.summarise.inform = FALSE)
 
+#' rename_values_eu
+#'
+#' Function to rename the codified values of the dataset to the meaningful
+#' values detailed in the mapping included in the package.
+#' @param data dataset to be standardized.
+#' @param current_var column name to be standardized.
+#' @importFrom dplyr %>%
+#' @return a dataset with labels renamed based in the mapping included in the package.
+#' @export
+rename_values = function(data, current_var) {
+  exchange_data = mapping %>%                                                # brings together the two df (mapping and data)
+    dplyr::filter(VAR_EN == current_var) %>%                                 # but only takes the part we are interested in (only when the value of the variable is equal to current_var).
+    dplyr::select(value, NAME) %>%
+    dplyr::distinct()
+
+  return(data)
+}
+
+
+#' standardize_eu
+#'
+#' Function to standarize data names.
+#' @param data dataset to be standardized.
+#' @param mapping mapping file to be used (included in medusa)
+#' @importFrom dplyr %>%
+#' @return a dataset with the variables and labels renamed based in the mapping included in the package.
+#' @export
+standardize <- function(data, mapping) {
+  # rename columns
+  old_names = colnames(data)                                                # creates a vector with the column names
+  new_names = dplyr::left_join(data.frame(VAR_EPF = old_names),             # the vector of column names is converted to a df and the column name is VAR_EPF
+                               mapping %>%
+                                 dplyr::select(VAR_EPF, VAR_EN) %>%         # within the mapping selects only the columns we are interested in
+                                 dplyr::distinct(.),                        # duplicate remove
+                               by = 'VAR_EPF') %>%                          # left join in function to var_epf
+    dplyr::mutate(VAR_EN = ifelse(is.na(VAR_EN), VAR_EPF, VAR_EN)) %>%      # when it is na you get var_epf and if it is not you get var
+    dplyr::pull(VAR_EN)                                                     # only keep the var column (as a vector)
+  colnames(data) = new_names                                                # assign to the column names the new vector
+
+  # rename values' codes to values' names for all items whose name is in the
+  # renamed mapping's column and have not NA values
+  ccitems = mapping %>%
+    dplyr::filter(VAR_EN %in% intersect(unique(mapping$VAR_EN), new_names),
+                  !is.na(value)) %>%
+    dplyr::pull(VAR_EN) %>%
+    unique()
+  for (cc in ccitems) {                                                     # for all columns where variables can be mapped we overwrite the standardised dataset (with rename_value function)
+    data = rename_values(data, cc)
+  }
+
+  return(data)
+}
+
 
 #' rawhbs_eu
 #'
@@ -150,3 +203,185 @@ rawhbs_eu <- function(year, country = "all", path) {
 }
 
 
+#' database_hbs
+#'
+#' Function to load each country HBS and create a joint database for household and member files
+#' @param year year/s of the HBS to process. Available options: 2010, 2015, 2020.
+#' @param country code of the country or countries of the HBS to process. By default, it processes all available data in the working directory To see the available countries and codes, run country_code().
+#' @param path Local path to the folder where the HBS's are stored. Not included in the package.
+#' @importFrom dplyr %>%
+#' @return RData file with the HBS microdata for each selected country and year.
[…]
