+++ R/utils_eu.R
@@ -3,6 +3,34 @@ library(magrittr)
 options(dplyr.summarise.inform = FALSE)
 
 
+#' rename_values_eu
+#'
+#' Function to rename the codified values of the dataset to the meaningful
+#' values detailed in the mapping included in the package.
+#' @param data dataset to be standardized.
+#' @param current_var column name to be standardized.
+#' @importFrom dplyr %>%
+#' @return a dataset with labels renamed based in the mapping included in the package.
+#' @export
+rename_values_eu = function(data, current_var) {
+  exchange_data = mapping %>%                                                # brings together the two df (mapping and data)
+    dplyr::filter(VAR_EN == current_var) %>%                                 # but only takes the part we are interested in (only when the value of the variable is equal to current_var).
+    dplyr::select(value, NAME) %>%
+    dplyr::distinct()
+
+  if (current_var != "NMIEMB" | (current_var == "NMIEMB" & sum(is.na(unique(exchange_data$value))) == 0) ) {                     # we look if there is any na in value and if so the below does not apply (to avoid errors with NMIEMB).
+    data = data %>%
+      dplyr::rename(value = {{ current_var }}) %>%                             # rename the column (variable name) to value
+      dplyr::mutate(value = as.character(value)) %>%                           # convert to a character
+      dplyr::left_join(exchange_data, by = "value") %>%                        # it keeps only the columns we are interested in (value and name).
+      dplyr::select(-value) %>%                                                # removes the value column
+      dplyr::rename_with(~current_var, 'NAME')                                 # rename to current_var
+  }
+
+
+  return(data)
+}
+
 #' standardize_eu
 #'
 #' Function to standarize data names.
@@ -31,7 +59,7 @@ standardize_eu <- function(data, mapping) {
     dplyr::pull(VAR_EN) %>%
     unique()
   for (cc in ccitems) {                                                     # for all columns where variables can be mapped we overwrite the standardised dataset (with rename_value function)
-    data = rename_values(data, cc)
+    data = rename_values_eu(data, cc)
   }
 
   return(data)
@@ -188,11 +216,11 @@ rawhbs_eu <- function(year, country = "all", path) {
 #' database_hbs
 #'
 #' Function to load each country HBS and create a joint database for household and member files
-#' @param year year/s of the HBS to process. Available options: 2010, 2015, 2020.
+#' @param year year of the HBS to process. Available options: 2015, 2020.
 #' @param country code of the country or countries of the HBS to process. By default, it processes all available data in the working directory To see the available countries and codes, run country_code().
-#' @param path Local path to the folder where the HBS's are stored. Not included in the package.
+#' @param inputs_path Local path to the folder where the HBS's are stored. Not included in the package.
 #' @importFrom dplyr %>%
-#' @return RData file with the HBS microdata for each selected country and year.
+#' @return RData file with the joint database with the HBS microdata.
 #' @export
 database_hbs <- function(year, country = "all", inputs_path) {
 
@@ -276,84 +304,80 @@ database_hbs <- function(year, country = "all", inputs_path) {
 
         load(file_path)
 
+        # Create income variable
+        Nquantiles <- function(x, w = NULL, s, t = NULL) {
+          if (is.null(t)) {
+            if (is.null(w)) w <- rep(1,length(x))
+
[…]
