### 2025-03-29 35aa81e
update database_hbs: to fix rename_values and merge reference person data
