+++ R/utils_eu.R
@@ -407,7 +407,7 @@ database_hbs <- function(year, country = "all", inputs_path) {
 
       # Create country2
       hbs_h <- hbs_h %>%
-        dplyr::mutate(country2 = dplyr::recode(country, "AT" =  "Austria"     ,
+        dplyr::mutate(country2 = dplyr::recode(COUNTRY, "AT" =  "Austria"     ,
                                                         "BE" =  "Belgium"     ,
                                                         "BG" =  "Bulgaria"    ,
                                                         "CY" =  "Cyprus"      ,
@@ -436,14 +436,12 @@ database_hbs <- function(year, country = "all", inputs_path) {
                                                         "SK" =  "Slovakia"    ))
 
       # Standardize household dataset variables --> fix rename values
-      hbs_h <- standardize_eu(hbs_h, variables_h)
+      #hbs_h <- standardize_eu(hbs_h, variables_h)
 
       # Save the modified data frame back to the list
       hbs_yearly_data[[name]] <- hbs_h
     }
   }
-
-
 
   # Code to create a joint database for 2020 and m file
   if (year == 2020) {
@@ -469,7 +467,7 @@ database_hbs <- function(year, country = "all", inputs_path) {
         # Rename values of education in SI to fix propblems with rename_values
         if (c == "SI") {
 
-          SI_hbs_m <- dplyr::mutate(SI_hbs_m, MC02B_agg = base::ifelse(MC02B_Recoded_3Categ ==  2, "A2",
+          data <- dplyr::mutate(data, MC02B_agg = base::ifelse(MC02B_Recoded_3Categ ==  2, "A2",
                                                           base::ifelse(MC02B_Recoded_3Categ ==  3, "A3",
                                                           base::ifelse(MC02B_Recoded_3Categ ==  5, "A5",
                                                           base::ifelse(MC02B_Recoded_3Categ == 88, "A88", NA)))))
@@ -515,6 +513,94 @@ database_hbs <- function(year, country = "all", inputs_path) {
 
   }
 
+  # Include reference person data from m file to h file --> FIX
+  if (year == 2020) {
+
+    # Extract hbs_h and hbs_m
+    hbs_h <- hbs_yearly_data$hbs_h_2020
+    hbs_m <- hbs_yearly_data$hbs_m_2020
+
+    # Subset m dataset: keep just data for the household reference person
+    hbs_m <- subset(hbs_m, relationship == "Reference person")
+
+    # CAUTION: nrows in hbs_m is higher than nrows in EU_hbs_h -> fix before merging
+    # Check duplicates
+    # for (c in country) {
+    #   h <- hbs_h %>% subset(country == c)
+    #   m <- hbs_m %>% subset(country == c)
+    #
+    #   nrows_m <- as.numeric(dplyr::count(h))
+    #
+    #   n_h <- length(unique(h$number))
+    #   n_m <- length(unique(m$number))
+    #   nurows_m <- as.numeric(n_m)
+    #
+    #   if (n_h != n_m) {
+    #     print(paste0("n_h & n_m don´t match in ", c))
+    #   }
+    #
+    #   if (nrows_m != nurows_m) {
+    #     print(paste0("unique values don´t match nrows in ", c))
+    #   }
+    #
+    #   h_n <- h[,c("number", "country")]
+    #   m_n <- m[,c("number", "country")]
+    #
+    #   dup_h <- sum(duplicated(h_n))
+    #   dup_m <- as.numeric(sum(duplicated(m_n)))
+    #
+    #   if (dup_m > 0) {
[…]
