+++ R/main.R
@@ -61,11 +61,10 @@ calc_di <-function(year, elevate=F, shocks, var_impact = "all", var_intersec = N
   # Check year parameter
   check_year(year)
 
-  if (!is.null(var_impact) && var_impact != "all") {
+  if (!is.null(var_impact) && length(var_impact) == 1 && var_impact != "all") {
     # Check var_impact
     check_var_impact(var_impact)
   }
-
 
   # get hbs files
   hbs <- get(paste0("epf_list_", year))
