+++ paper/paper.bib
@@ -207,6 +207,7 @@
   title={Some contributions of integrated assessment models of global climate change},
   author={Weyant, John},
   journal={Review of Environmental Economics and Policy},
+  doi = {10.1093/reep/rew018},
   year={2017},
   publisher={The University of Chicago Press}
 }
