+++ R/utils.R
@@ -784,7 +784,7 @@ adjust_wh_is <- function(data, var_w, var_h) {
 
 #' order_var
 #'
-#' Function to order the labels of the socioeconomic and demographic variables
+#' Function to order the labels of the socioeconomic and demographic variablesin basic graph
 #' @param data dataset in which we want to order the labels of the socioeconomic and demographic variables
 #' @param g variable for which we want to sort the labels
 #' @importFrom dplyr %>%
@@ -793,22 +793,22 @@ adjust_wh_is <- function(data, var_w, var_h) {
 order_var <- function(data, g){
   if (g == "CHILDREN"){
     data <- data %>%
-    dplyr::mutate(CHILDREN = factor(CHILDREN, levels = c("No children", "With children", "Large family")))
+    dplyr::mutate(LABELS = factor(LABELS, levels = c("No children", "With children", "Large family")))
   } else if (g == "COUNTRYRP") {
     data <- data %>%
-    dplyr::mutate(COUNTRYRP = factor(COUNTRYRP, levels = c("Spain", "EU27", "Other Europe", "Rest of world")))
+    dplyr::mutate(LABELS = factor(LABELS, levels = c("Spain", "EU27", "Other Europe", "Rest of world")))
   } else if (g == "STUDIESRP") {
     data <- data %>%
-    dplyr::mutate(STUDIESRP = factor(STUDIESRP, levels = c("Without studies", "Primary education", "Secondary education", "Post-secondary education", "Higher education")))
+    dplyr::mutate(LABELS = factor(LABELS, levels = c("Without studies", "Primary education", "Secondary education", "Post-secondary education", "Higher education")))
   } else if (g == "REGMR") {
     data <- data %>%
-    dplyr::mutate(REGMR = factor(REGMR, levels = c("Rented","Ownership", "Relinquish")))
+    dplyr::mutate(LABELS = factor(LABELS, levels = c("Rented","Ownership", "Relinquish")))
   } else if (g == "PROFESSIONALSRP") {
     data <- data %>%
-    dplyr::mutate(PROFESSIONALSRP = factor(PROFESSIONALSRP, levels = c("Employee", "Self-employed", "Employer")))
+    dplyr::mutate(LABELS = factor(LABELS, levels = c("Employee", "Self-employed", "Employer")))
   } else if (g == "AGERP") {
     data <- data %>%
-      dplyr::mutate(AGERP = factor(AGERP, levels = c("Young", "Adult", "Elder")))
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("Young", "Adult", "Elder")))
   }
   return(data)
 }
@@ -833,7 +833,7 @@ basic_graph <- function(data, var = categories$categories){
       tidyr::pivot_longer(cols = dplyr::starts_with("DI_"), names_to = "Scenario", values_to = "Impact") %>%
       dplyr::mutate(Scenario = stringr::str_replace(Scenario, "^DI_", "")) %>%
       dplyr::filter(! LABELS %in% c("Not provided", "NA", "Others")) %>%
-      order_var(., "LABELS")
+      order_var(., g)
 
     # Convert LABELS to character if not DECILE
     if (g == "DECILE") {
@@ -944,6 +944,37 @@ impact <- function(data, var = categories$categories, save = T, file_name = "D_i
 
 }
 
+#' order_vars
+#'
+#' Function to order the labels of the socioeconomic and demographic variables in intersectional graph
+#' @param data dataset in which we want to order the labels of the socioeconomic and demographic variables
+#' @param g variable for which we want to sort the labels
+#' @importFrom dplyr %>%
+#' @return a dataset in which the labels are ordered for the selected socioeconomic or demographic variable
+#' @export
+order_var <- function(data, g){
+  if (g == "CHILDREN"){
+    data <- data %>%
[…]
