+++ R/utils_eu.R
@@ -8,12 +8,13 @@ options(dplyr.summarise.inform = FALSE)
 #' Function to rename the codified values of the dataset to the meaningful
 #' values detailed in the mapping included in the package.
 #' @param data dataset to be standardized.
+#' @param map mapping file to be used (included in medusa)
 #' @param current_var column name to be standardized.
 #' @importFrom dplyr %>%
 #' @return a dataset with labels renamed based in the mapping included in the package.
 #' @export
-rename_values_eu = function(data, current_var) {
-  exchange_data = mapping %>%                                                # brings together the two df (mapping and data)
+rename_values_eu = function(data, map = variables_h, current_var) {
+  exchange_data = map %>%                                                # brings together the two df (mapping and data)
     dplyr::filter(VAR_EN == current_var) %>%                                 # but only takes the part we are interested in (only when the value of the variable is equal to current_var).
     dplyr::select(value, NAME) %>%
     dplyr::distinct()
@@ -23,6 +24,7 @@ rename_values_eu = function(data, current_var) {
       dplyr::rename(value = {{ current_var }}) %>%                             # rename the column (variable name) to value
       dplyr::mutate(value = as.character(value)) %>%                           # convert to a character
       dplyr::left_join(exchange_data, by = "value") %>%                        # it keeps only the columns we are interested in (value and name).
+      dplyr::mutate(NAME = dplyr::if_else(is.na(NAME), value, NAME)) %>%
       dplyr::select(-value) %>%                                                # removes the value column
       dplyr::rename_with(~current_var, 'NAME')                                 # rename to current_var
   }
@@ -35,15 +37,15 @@ rename_values_eu = function(data, current_var) {
 #'
 #' Function to standarize data names.
 #' @param data dataset to be standardized.
-#' @param mapping mapping file to be used (included in medusa)
+#' @param map mapping file to be used (included in medusa)
 #' @importFrom dplyr %>%
 #' @return a dataset with the variables and labels renamed based in the mapping included in the package.
 #' @export
-standardize_eu <- function(data, mapping) {
+standardize_eu <- function(data, map = variables_h) {
   # rename columns
   old_names = colnames(data)                                                # creates a vector with the column names
   new_names = dplyr::left_join(data.frame(VAR_EPF = old_names),             # the vector of column names is converted to a df and the column name is VAR_EPF
-                               mapping %>%
+                               map %>%
                                  dplyr::select(VAR_EPF, VAR_EN) %>%         # within the mapping selects only the columns we are interested in
                                  dplyr::distinct(.),                        # duplicate remove
                                by = 'VAR_EPF') %>%                          # left join in function to var_epf
@@ -53,13 +55,13 @@ standardize_eu <- function(data, mapping) {
 
   # rename values' codes to values' names for all items whose name is in the
   # renamed mapping's column and have not NA values
-  ccitems = mapping %>%
-    dplyr::filter(VAR_EN %in% intersect(unique(mapping$VAR_EN), new_names),
+  ccitems = map %>%
+    dplyr::filter(VAR_EN %in% intersect(unique(map$VAR_EN), new_names),
[…]
