### 2025-04-20 8401c01
add price_shock_eu
