### 2025-04-20 86f52ad
add price_shock_eu for multiple scenarios
