### 2025-04-21 aaceabc
modify impact_eu to add by_country
