+++ R/utils_eu.R
@@ -873,75 +873,6 @@ update_year <- function(data, new_year){
 #' @export
 price_shock_eu <- function(data, shocks) {
 
-  # Obtener la lista de países (intersección entre columnas de shocks y valores en data$country)
-  countries <- unique(data$country)
-  common_countries <- intersect(countries, colnames(shocks))
-
-  # Identificar columnas COICOP (todas las que empiezan por CP y no son CP00)
-  coicop_cols <- grep("^CP\\d+", colnames(data), value = TRUE)
-  coicop_cols <- setdiff(coicop_cols, "CP00")
-
-  for (country in common_countries) {
-
-    # Filas del país actual
-    idx <- which(data$country == country)
-
-    for (col in coicop_cols) {
-
-      # Shock correspondiente para el código coicop y país
-      shock_val <- shocks[shocks$coicop == col, country]
-
-      # Aplicar solo si shock distinto de 1 y no es NA
-      if (length(shock_val) == 1 && !is.na(shock_val) && shock_val != 1) {
-
-        # Nombre nueva columna
-        new_col <- paste0(col, "_new")
-
-        # Si la columna nueva aún no existe, crearla como copia del original
-        if (!(new_col %in% colnames(data))) {
-          data[[new_col]] <- data[[col]]
-        }
-
-        # Aplicar solo si shock distinto de 1 y no es NA
-        if (length(shock_val) == 1 && !is.na(shock_val) && shock_val != 1) {
-          data[idx, new_col] <- data[idx, col] * shock_val
-        }
-      }
-    }
-
-    # Calcular la nueva CP00 como original + diferencia con columnas modificadas
-    diff_cols <- grep("_new$", names(data), value = TRUE)
-    orig_cols <- gsub("_new$", "", diff_cols)
-
-    if (length(diff_cols) == 1) {
-      # Si solo hay una columna afectada
-      diff_val <- data[idx, diff_cols] - data[idx, orig_cols]
-    } else {
-      # Si hay más de una, tratamos como data.frame
-      diff_val <- data[idx, diff_cols] - data[idx, orig_cols]
-    }
-
-    # Asegurarse de que diff_val tenga dimensión de matriz
-    if (is.null(dim(diff_val))) {
-      diff_val <- matrix(diff_val, ncol = 1)
-    }
-
-    # Asegurar que CP00_new existe
-    if (!"CP00_new" %in% colnames(data)) {
-      data$CP00_new <- data$CP00
-    }
-
-    # Calcular nueva CP00
-    data$CP00_new[idx] <- data$CP00[idx] + rowSums(diff_val, na.rm = TRUE)
-  }
-
-  return(data)
-}
-
-
-### PRUEBA VARIOS ESCENARIOS
-price_shock_eu <- function(data, shocks) {
-
   # Extraer países presentes en los datos
   countries <- unique(data$country)
 
@@ -1002,4 +933,134 @@ price_shock_eu <- function(data, shocks) {
 
   return(data)
 }
+
+
+#' impact_eu
+#'
+#' Function to calculate the distributional impacts based in one or more
+#' socioeconomic or demographic variables (one impact per variable) for EU countries.
+#' @param data a dataset with the input data needed to calculate distributional
+#' impacts. The dataset should contain both the household expenditures collected
+#' in the HBS and the expenditures after applying the price shock.
+#' @param var variable(s) according to which you want to calculate distributional
+#' impacts. If categories$categories (by default) calculates the distributional
+#' impacts for each of the variables specified in the package. If not, you can
+#' indicate a variable or a vector of variables to calculate distributional impacts.
[…]
