### 2025-04-21 1286739
add basic_graph_eu & order_var_eu
