+++ R/utils_eu.R
@@ -935,6 +935,145 @@ price_shock_eu <- function(data, shocks) {
 }
 
 
+#' order_var_eu
+#'
+#' Function to order the labels of the socioeconomic and demographic variables
+#' @param data dataset in which we want to order the labels of the socioeconomic and demographic variables
+#' @param g variable for which we want to sort the labels
+#' @importFrom dplyr %>%
+#' @return a dataset in which the labels are ordered for the selected socioeconomic or demographic variable
+#' @export
+order_var_eu <- function(data, g){
+  if (g == "children"){
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("No children", "With children", "Large family")))
+  } else if (g == "birth_country") {
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("National", "EU", "Non-EU", "Non-national")))
+  } else if (g == "education") {
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("Early childhood education", "Primary education", "Secondary education", "Post-secondary education",  "Tertiary education", "Higher education")))
+  } else if (g == "contract_type") {
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("Permanent contract", "Fixed-term contract", "Not applicable")))
+  } else if (g == "activity") {
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("Employed", "Unemployed", "Retired", "Student", "Domestic tasks", "Disabled", "Military service")))
+  } else if (g == "age") {
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("Young", "Adult", "Elder")))
+  } else if (g == "employment_sector") {
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("Private sector", "Public sector")))
+  } else if (g == "household_type") {
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c("Adult alone", "Couple", "Couple with children", "More than 2 adults", "More than 2 adults with children", "Single parent"  )))
+  } else if (g == "income_source") {
+    data <- data %>%
+      dplyr::mutate(LABELS = factor(LABELS, levels = c( "Wages", "Self-employment", "Property", "Pensions", "Unemployment")))
+  }
+  return(data)
+}
+
+#' basic_graph_eu
+#'
+#' Function to create a basic graph to summarize the distributional impact in the EU
+#' based in one or more socioeconomic or demographic variable (one plot per variable).
+#' @param data a dataset with the input data needed to generate a basic graph.
+#' @param var variable(s) according to which you want to generate the graph. If
+#' graph_labels_eu$VARIABLE (by default) creates a graph with the distributional
+#' impacts for each of the variables specified in the package. If not, you can
+#' indicate a variable or a vector of variables to crate the graph.If you want to
+#' see the variables for which the function is available run `available_var_impact_eu()`.
+#' @importFrom dplyr %>%
+#' @return a graph per selected variable/s summarizing distributional impacts.
+#' @export
+basic_graph_eu <- function(data, var = graph_labels_eu$VARIABLE) {
+  if (!dir.exists("figures")) dir.create("figures")
+
+  # Si es una lista de dataframes (como en el output original de impact_eu)
[…]
