+++ inst/extdata/saveDataFiles.R
@@ -198,6 +198,10 @@ use_data(shocks_eu, overwrite=T)
 shocks_eu_s = read.csv(paste0(rawDataFolder,"shocks_eu_s.csv"),header=T)
 use_data(shocks_eu_s, overwrite=T)
 
-# graph_labels
+# graph_labels_eu
 graph_labels_eu = read.csv(paste0(rawDataFolder,"graph_labels_eu.csv"),header=T)
 use_data(graph_labels_eu, overwrite=T)
+
+# is_categories_eu
+is_categories_eu = read.csv(paste0(rawDataFolder,"is_categories_eu.csv"),header=T)
+use_data(is_categories_eu, overwrite=T)
