### 2025-04-22 #7697 @thegargiulian
Hi @evaaepelde, thank you!

The edits to the paper, README, and data documentation look great and I've update my checklist accordingly! Unfortunately, after installing the new version of `medusa` I still can't get the `calc_di` example with `vars <- c("ZONE", "QUINTILE")` to work and I'm getting the same error as before (from this tutorial page: https://bc3lc.github.io/medusa/articles/Tutorials.html)
