+++ R/utils_eu.R
@@ -1061,7 +1061,7 @@ basic_graph_eu <- function(data, var = graph_labels_eu$VARIABLE) {
 
       # Aumenta anchura si g == "country"
       if (g == "country") {
-        adj_wh$width <- adj_wh$width + 50  # Puedes ajustar el valor (p.ej., +70 si aún es pequeño)
+        adj_wh$width <- adj_wh$width + 50
       }
 
       ggplot2::ggsave(pl,
@@ -1203,6 +1203,137 @@ impact_eu <- function(data, var = graph_labels_eu$VARIABLE, by_country = TRUE, s
   return(combined_df)
 }
 
+
+#' order_vars_eu
+#'
+#' Function to order the labels of the socioeconomic and demographic variables in intersectional graphs.
+#' @param data dataset in which we want to order the labels of the socioeconomic and demographic variables
+#' @param g variable for which we want to sort the labels
+#' @importFrom dplyr %>%
+#' @return a dataset in which the labels are ordered for the selected socioeconomic or demographic variable
+#' @export
+order_vars_eu <- function(data, var_col, col_name = "LABELS") {
+  levels_list <- list(
+    children = c("No children", "With children", "Large family"),
+    birth_country = c("National", "EU", "Non-EU", "Non-national"),
+    education = c("Early childhood education", "Primary education", "Secondary education",
+                  "Post-secondary education", "Tertiary education", "Higher education"),
+    contract_type = c("Permanent contract", "Fixed-term contract", "Not applicable"),
+    activity = c("Employed", "Unemployed", "Retired", "Student", "Domestic tasks", "Disabled", "Military service"),
+    age = c("Young", "Adult", "Elder"),
+    employment_sector = c("Private sector", "Public sector"),
+    household_type = c("Adult alone", "Couple", "Couple with children", "More than 2 adults",
+                       "More than 2 adults with children", "Single parent"),
+    income_source = c("Wages", "Self-employment", "Property", "Pensions", "Unemployment")
+  )
+
+  if (var_col %in% names(levels_list)) {
+    lvls <- levels_list[[var_col]]
+    data[[col_name]] <- factor(data[[col_name]], levels = lvls)
+  }
+
+  return(data)
+}
+
+
+#' intersectional_graph_eu
+#'
+#' Function to create an intersectional graph to summarize the distributional
+#' impact based in the intersection of two socioeconomic or demographic variables
+#' (2 variables per plot) for EU countries.
+#' @param data a dataset with the input data needed to generate the intersectional graph.
+#' @param pairs set of variables (2) according to which you want to create the
+#' intersectional graph. If is_categories (by default), it generates the intersectional
+#' graph for each of the combinations of variables specified in the package. If not,
+#' you can indicate the set of variables according to which you want to generste the
+#' intersectional graph. If you wish to see the set of variables for which the
+#' calculation is available, run `available_var_intersec_eu()`. To enter a set of
+#' variables for the calculation, it must follow the same format as the output of
+#' `available_var_intersec_eu()`, i.e. a table whose columns have category_a and
+#' category_b as their titles.
+#' @importFrom dplyr %>%
[…]
