### 2025-05-07 08a8c87
GCAM: get_gcam_prices script
