+++ inst/extdata/saveDataFiles.R
@@ -209,3 +209,100 @@ use_data(graph_labels_eu, overwrite=T)
 # is_categories_eu
 is_categories_eu = read.csv(paste0(rawDataFolder,"is_categories_eu.csv"),header=T)
 use_data(is_categories_eu, overwrite=T)
+# mapping GCAM - MEDUSA COICOP categories
+mapping_gcam_medusa <- read.csv(paste0(rawDataFolder, "mapping_gcam_medusa.csv"), header = T)
+use_data(mapping_gcam_medusa, overwrite=T)
+
+# coicop codes list
+all_coicop <- c(
+  "CP01", "CP011", "CP0111", "CP01111", "CP01112", "CP01113", "CP01114", "CP01115", "CP01116", "CP01117", "CP01118",
+  "CP0112", "CP01121", "CP01122", "CP01123", "CP01124", "CP01125", "CP01126", "CP01127", "CP01128",
+  "CP0113", "CP01131", "CP01132", "CP01133", "CP01134", "CP01135", "CP01136",
+  "CP0114", "CP01141", "CP01142", "CP01143", "CP01144", "CP01145", "CP01146", "CP01147",
+  "CP0115", "CP01151", "CP01152", "CP01153", "CP01154", "CP01155",
+  "CP0116", "CP01161", "CP01162", "CP01163", "CP01164",
+  "CP0117", "CP01171", "CP01172", "CP01173", "CP01174", "CP01175", "CP01176",
+  "CP0118", "CP01181", "CP01182", "CP01183", "CP01184", "CP01185", "CP01186",
+  "CP0119", "CP01191", "CP01192", "CP01193", "CP01194", "CP01199",
+  "CP012", "CP0121", "CP01211", "CP01212", "CP01213",
+  "CP0122", "CP01221", "CP01222", "CP01223",
+  "CP02", "CP021", "CP0211", "CP02111", "CP02112",
+  "CP0212", "CP02121", "CP02122", "CP02123", "CP02124",
+  "CP0213", "CP02131", "CP02132", "CP02133", "CP02134",
+  "CP022", "CP0220", "CP02201", "CP02202", "CP02203",
+  "CP023",
+  "CP03", "CP031", "CP0311", "CP03110", "CP0312", "CP03121", "CP03122", "CP03123",
+  "CP0313", "CP03131", "CP03132",
+  "CP0314", "CP03141", "CP03142",
+  "CP032", "CP0321", "CP03211", "CP03212", "CP03213",
+  "CP0322", "CP03220",
+  "CP04", "CP041", "CP0411", "CP04110", "CP0412", "CP04121", "CP04122",
+  "CP042", "CP0421", "CP04210", "CP0422", "CP04220",
+  "CP043", "CP0431", "CP04310", "CP0432", "CP04321", "CP04322", "CP04323", "CP04324", "CP04325", "CP04329",
+  "CP044", "CP0441", "CP04410", "CP0442", "CP04420", "CP0443", "CP04430", "CP0444", "CP04441", "CP04442", "CP04449",
+  "CP045", "CP0451", "CP04510", "CP0452", "CP04521", "CP04522", "CP0453", "CP04530", "CP0454", "CP04541", "CP04549",
+  "CP0455", "CP04550",
+  "CP05", "CP051", "CP0511", "CP05111", "CP05112", "CP05113", "CP05119",
+  "CP0512", "CP05121", "CP05122", "CP05123", "CP0513", "CP05130",
+  "CP052", "CP0520", "CP05201", "CP05202", "CP05203", "CP05204", "CP05209",
[…]
