### 2025-05-07 87701a3
GCAM: get_gcam_prices script
