+++ R/get_gcam_prices.R
@@ -1,0 +1,214 @@
+# library(tidyr)
+# library(dplyr)
+# library(magrittr)
+# library(ggplot2)
+# library(rgcam)
+# library(gcamdata)
+# library(countrycode)
+# library(tibble)
+
+#------------------------
+#' get_prices_gcam
+#'
+#' Extract and format prices from a GCAM-Europe database or project file for MEDUSA
+#' @keywords GCAM, GCAM-Europe, prices
+#' @return a dataframe with prices by sector and GCAM-Europe region
+#' @param db_path Path to the GCAM database
+#' @param query_path Path to the query file
+#' @param db_name Name of the GCAM database
+#' @param prj_name Name of the rgcam project. This can be an existing project, or, if not, this will be the name
+#' @param prj rgcam loaded project
+#' @param scenarios Vector names of the GCAM scenarios to be processed
+#' @param queries Name of the GCAM query file. The file by default includes the queries required to run rfasst
+#' @param final_db_year Final year in the GCAM database (this allows to process databases with user-defined "stop periods")
+#' @param saveOutput Writes the files.By default=T
+#' @param base_scen The base scenario that other scenarios will be compared against
+#' @param selected_year The year of analysis, when the scenarios wil be compared
+#' @importFrom magrittr %>%
+#' @export
+get_prices_gcam <- function(db_path = NULL, query_path = "inst/extdata", db_name = NULL, prj_name, prj = NULL,
+                            scenarios, queries = "queries_GCAM_MEDUSA.xml", final_db_year = 2100,
+                            saveOutput = T, base_scen, selected_year) {
+
+  db_path <- 'C:\\Users\\claudia.rodes\\Documents\\GitHub\\gcam-core-eu\\output'
+  db_name <- 'database_basexdb_heatpumps2'
+  query_path = "inst/extdata"
+  prj_name = 'medusa.dat'
+  prj = NULL
+  scenarios = c('Reference_HeatPumps17_2100','Reference_heatpumps')
+  queries = "queries_GCAM_MEDUSA.xml"
+  final_db_year = 2030
+  saveOutput = T
+  base_scen = 'Reference_heatpumps'
+  selected_year = 2015
+
+
+  # Set countries
+  EU_COUNTRIES <- c("Austria", "Belgium", "Bulgaria", "Croatia",
+                    "Cyprus", "Czech Republic", "Denmark",
+                    "Estonia", "Finland", "France", "Germany",
+                    "Greece", "Hungary", "Ireland", "Italy",
+                    "Latvia", "Lithuania", "Luxembourg", "Malta",
+                    "Netherlands", "Poland", "Portugal", "Romania",
+                    "Slovakia", "Slovenia", "Spain", "Sweden")
+
+  crop_prod_COUNTRY <- "Austria"
+
+  # Set crop coicop codes (for adjustment)
+  crop_coicop_codes <- c("CP01111",
+                         "CP01112",
+                         "CP01154",
+                         "CP0116",
+                         "CP01163",
+                         "CP0117",
+                         "CP01176",
+                         "CP01181")
+
+  # Set delivered biomass code (not working well)
+  deliv_bio_coicop <- "CP04549"
+
+  # Load price mapping files
+  map_price <- medusa::mapping_gcam_medusa %>%
+    dplyr::filter(query == "prices by sector") %>%
+    dplyr::select(sector = gcam) %>%
+    dplyr::distinct() %>%
+    dplyr::pull()
+
[…]
