+++ R/data.R
@@ -662,3 +662,15 @@
 #' @source medusa
 #' @format .csv
 "graph_labels"
+
+#' Mapping between the GCAM items names and the COICOP codes
+#'
+#' @source medusa
+#' @format .csv
+"mapping_gcam_medusa"
+
+#' List of all COICOP codes available in MEDUSA
+#'
+#' @source medusa
+#' @format .csv
+"all_coicop"
