+++ man/order_vars.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/utils.R
-\name{order_var}
-\alias{order_var}
-\title{order_var}
+\name{order_vars}
+\alias{order_vars}
+\title{order_vars}
 \usage{
-order_var(data, g)
+order_vars(data, g)
 }
 \arguments{
 \item{data}{dataset in which we want to order the labels of the socioeconomic and demographic variables}
@@ -15,5 +15,5 @@ order_var(data, g)
 a dataset in which the labels are ordered for the selected socioeconomic or demographic variable
 }
 \description{
-Function to order the labels of the socioeconomic and demographic variablesin basic graph
+Function to order the labels of the socioeconomic and demographic variables in intersectional graph
 }
