+++ man/mapping_gcam_medusa.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{mapping_gcam_medusa}
+\alias{mapping_gcam_medusa}
+\title{Mapping between the GCAM items names and the COICOP codes}
+\format{
+.csv
+}
+\source{
+medusa
+}
+\usage{
+mapping_gcam_medusa
+}
+\description{
+Mapping between the GCAM items names and the COICOP codes
+}
+\keyword{datasets}
