+++ NAMESPACE
@@ -14,6 +14,7 @@ export(check_year)
 export(elevate_hbs)
 export(ex_shocks)
 export(ex_var_intersec)
+export(get_prices_gcam)
 export(id_ep1)
 export(id_ep2)
 export(id_tp)
@@ -22,9 +23,11 @@ export(impact_intersectional)
 export(intersectional_graph)
 export(load_rawhbs)
 export(order_var)
+export(order_vars)
 export(price_shock)
 export(rename_values)
 export(standardize)
 export(weighted.median)
 export(weighted.quantile)
 importFrom(dplyr,"%>%")
+importFrom(magrittr,"%>%")
