### 2025-05-07 #7697 @thegargiulian
Thanks @evaaepelde!

In case it's helpful, I tested the example on MacOS with R version 4.4.0.
