### 2025-05-12 b52dc17
add hbs_eu and start calc_di_eu
