+++ R/main_eu.R
@@ -23,13 +23,13 @@ ex_shocks_eu <- function(){
 }
 
 
-#' available_var_impact_eu
+#' available_var_eu
 #'
 #' Function that returns the variables for which the calculation of
 #' basic distributional impacts is available for EU countries.
 #' @return variables for which the calculation of basic distributional impacts is available
 #' @export
-available_var_impact_eu <- function(){
+available_var_eu <- function(){
   av_var_impact <- graph_labels_eu$VARIABLE
   print(av_var_impact)
 }
@@ -45,4 +45,163 @@ available_var_impact_eu <- function(){
 available_var_intersec <- function(){
   av_var_intersec <- is_categories_eu
   print(av_var_intersec)
+}
+
+
+#' hbs_eu
+#'
+#' Main function to generate the database (from EUROSTAT microdata) to calculate
+#' the distributional impacts with calc_di_eu function.
+#' @param year year of the HBS to process. Available options: 2010, 2015, 2020.
+#' @param country code of the country or countries of the HBS to process and calculate
+#' distributional impacts. By default, it processes all available data in the working
+#' directory To see the available countries and codes, run country_code().
+#' @param path Local path to the folder where the HBS's are stored (microdata from
+#' EUROSTAT, not included in the package). In the defined folder, the data provided
+#' by EUROSTAT must be saved in a folder with the name of the year to which they belong.
+#' @importFrom dplyr %>%
+#' @return a list containing the generated datasets summarising the basic or/and the
+#' intersectional distributional impacts per selected variable or set of variables.
+#' @export
+hbs_eu <- function(year, country = "all", path) {
+
+  # Process raw data from EUROSAT
+  rawhbs_eu(year, country = "all", path)
+
+  # Create a joint database for household and member files
+  if (length(year) != 1) {
+    # Check that only one year is provided
+    stop("Error: You have provided a vector of years to the 'year' argument. This function only works with a single year.")
+  } else {
+    inputs_path <- file.path(path, "inputs")
+    hbs <- database_hbs(year, country, inputs_path)
+    hbs <- hbs[[paste0("hbs_", year)]]
+  }
+
+  # Rename expenditure variables
+  hbs <- rename_coicop(data = hbs)
+
+  return(hbs)
+
+}
+
+
+#' calc_di_eu
+#'
+#' Main function to calculate the distributional impacts for different price shocks
+#' and to save the results and the figures.
+#' @param data Local
+#' @param update_hbs year to which you want to update the microdata. Include only
+#' if you want to update the HBS expenditure data to a year other than the base
+#' year. By default, the parameter is F, so skip this step.
+#' @param shocks a dataset with the price shocks per coicop to be applied.
+#' The format of the dataset has to correspond to the predefined one in the package.
+#' To save a csv file with the right format to enter the price shocks run `ex_shocks_eu()`.
+#' You can enter more scenarios by including more columns to the right with all
+#' the country codes + the scenario name (e.g. AT_s3, BE_s3...). A price shock
+#' greater than 1 indicates a price increase (e.g. 1.1 indicates a 10\% increase)
+#' and less than 1 indicates a price decrease (e.g. 0.9 indicates a 10\% decrease).
+#' @param outputs_path Local path to the folder where the results will be stored
+#' (not included in the package). In the defined folder, the data provided
+#' by EUROSTAT must be saved in a folder with the name of the year to which they belong.
+#' By default, it creates and saves the results in the outputs folder within your working directory.
+#' @param var variable(s) according to which you want to calculate distributional
+#' impacts. If "all" (by default) calculates the distributional impacts for each of the
+#' variables specified in the package.If not, you can indicate a variable or a vector
[…]
