### 2025-05-13 16a615f
update calc_di_eu
