+++ R/main_eu.R
@@ -42,7 +42,7 @@ available_var_eu <- function(){
 #' @return set of variables for which the calculation of intersectional
 #' distributional impacts is available for EU countries
 #' @export
-available_var_intersec <- function(){
+available_var_intersec_eu <- function(){
   av_var_intersec <- is_categories_eu
   print(av_var_intersec)
 }
@@ -102,9 +102,8 @@ hbs_eu <- function(year, country = "all", path) {
 #' greater than 1 indicates a price increase (e.g. 1.1 indicates a 10\% increase)
 #' and less than 1 indicates a price decrease (e.g. 0.9 indicates a 10\% decrease).
 #' @param outputs_path Local path to the folder where the results will be stored
-#' (not included in the package). In the defined folder, the data provided
-#' by EUROSTAT must be saved in a folder with the name of the year to which they belong.
-#' By default, it creates and saves the results in the outputs folder within your working directory.
+#' (not included in the package). By default, it creates and saves the results in
+#' the outputs folder within your working directory.
 #' @param var variable(s) according to which you want to calculate distributional
 #' impacts. If "all" (by default) calculates the distributional impacts for each of the
 #' variables specified in the package.If not, you can indicate a variable or a vector
@@ -139,18 +138,18 @@ calc_di_eu <-function(data, update_hbs=F, shocks, outputs_path=F, var = "all",
                       var_intersec = NULL, by_country = TRUE, save=T, fig=TRUE,
                       file_name = "D_impact", file_name_intersec = "DI_impact") {
 
-
   # Update hbs
   if(update_hbs != F) {
-    hbs <- update_year(data, new_year = update_hbs )
+    data <- update_year(data, new_year = update_hbs )
   }
 
   # Apply price shocks
-  hbs <- price_shock_eu(data = hbs, shocks)
+  hbs <- price_shock_eu(data, shocks)
 
   # Set working directory to save results
   if (outputs_path == F) {
     outputs_path <- file.path(path, "outputs")
+    if (!dir.exists(outputs_path)) dir.create(outputs_path, recursive = TRUE)
     setwd(outputs_path)
   } else {
     setwd(outputs_path)
@@ -194,7 +193,7 @@ calc_di_eu <-function(data, update_hbs=F, shocks, outputs_path=F, var = "all",
   # create a list to return
   return_list <- list()
 
-  if(!is.null(var_impact)){
+  if(!is.null(var)){
     return_list[["di"]] <- di
   }
 
