### 2025-05-13 7eabd48
update calc_di_eu
