### 2025-05-13 2bd52dc
update calc_di_eu to add intersectional graphs by country
