+++ R/utils_eu.R
@@ -1267,85 +1267,100 @@ intersectional_graph_eu <- function(data, pairs = is_categories_eu) {
   for (r in 1:nrow(pairs)) {
     var_a <- pairs$category_a[r]
     var_b <- pairs$category_b[r]
-    df <- data[[paste0("di_", var_a, "_", var_b)]]
-    if (is.null(df)) next
 
-    source_tag <- unique(df$SOURCE)
-    folder <- ifelse(length(source_tag) == 1 && source_tag %in% c("EU"), "EU", source_tag)
-    folder_path <- file.path("figures", folder)
-    if (!dir.exists(folder_path)) dir.create(folder_path)
+    # Buscar todos los objetos que coincidan con este par de variables
+    matching_keys <- grep(paste0("^di_", var_a, "_", var_b), names(data), value = TRUE)
 
-    datapl <- df %>%
-      tidyr::pivot_longer(cols = dplyr::starts_with("DI_"), names_to = "Scenario", values_to = "Impact") %>%
-      dplyr::mutate(
-        Scenario = stringr::str_replace(Scenario, "^DI_", ""),
-        LABELS_A = as.character(LABELS_A),
-        LABELS_B = as.character(LABELS_B)
-      )
+    for (key in matching_keys) {
+      df <- data[[key]]
+      if (is.null(df)) next
 
-    # Ordenar
-    datapl <- order_vars_eu(datapl, var_a, "LABELS_A")
-    datapl <- order_vars_eu(datapl, var_b, "LABELS_B")
+      source_tag <- unique(df$SOURCE)
+      folder <- ifelse(length(source_tag) == 1 && source_tag %in% c("EU"), "EU", source_tag)
+      folder_path <- file.path("figures", folder)
+      if (!dir.exists(folder_path)) dir.create(folder_path)
 
-    datapl <- datapl %>%
-      dplyr::filter(
-        !LABELS_A %in% c("Not provided", "Others", "Other", "Not applicable"),
-        !LABELS_B %in% c("Not provided", "Others", "Other", "Not applicable")
-      ) %>%
-      dplyr::filter(!is.na(LABELS_A) & LABELS_A != "NA") %>%
-      dplyr::filter(!is.na(LABELS_B) & LABELS_B != "NA") %>%
-      droplevels()
+      datapl <- df %>%
+        tidyr::pivot_longer(cols = dplyr::starts_with("DI_"), names_to = "Scenario", values_to = "Impact") %>%
+        dplyr::mutate(
+          Scenario = stringr::str_replace(Scenario, "^DI_", ""),
+          LABELS_A = as.character(LABELS_A),
+          LABELS_B = as.character(LABELS_B)
+        )
 
-    clean_a <- graph_labels_eu %>% dplyr::filter(VARIABLE == var_a) %>% dplyr::pull(VAR_CLEAN)
-    clean_b <- graph_labels_eu %>% dplyr::filter(VARIABLE == var_b) %>% dplyr::pull(VAR_CLEAN)
+      # Ordenar
+      datapl <- order_vars_eu(datapl, var_a, "LABELS_A")
+      datapl <- order_vars_eu(datapl, var_b, "LABELS_B")
 
-    if (var_a %in% c("decile", "decile_eu", "quintile", "quintile_eu", "ventile", "ventile_eu", "percentile", "percentile_eu")) {
-      pl <- ggplot2::ggplot(datapl, ggplot2::aes(x = as.numeric(LABELS_A), y = Impact, colour = Scenario)) +
-        ggplot2::geom_point() +
-        ggplot2::geom_line() +
-        ggplot2::facet_grid(LABELS_B ~ Scenario) +
[…]
