### 2025-05-14 508dfeb
update intersectional_graph_eu to skip plots with insuficient ata
