### 2025-05-19 #7697 @evaaepelde
Hi @thegargiulian, thank you again!

We've been testing the example on several different machines and operating systems, and so far it seems to work correctly. However, we haven’t been able to test it on a Mac, as we currently don’t have access to one, so that might explain why we haven’t been able to reproduce the issue.

Just to double-check: could you please make sure you’re using the latest version of the medusa package? Sometimes updating via install.packages() doesn’t overwrite properly, so one workaround could be to clone the GitHub repository directly and install from there: remotes::install_github("BC3LC/medusa")

Let us know if that helps or if the issue persists — we’re keen to get to the bottom of it and really appreciate your efforts in testing the package so thoroughly!
