### 2025-05-25 #7697 @thegargiulian
After uninstalling and re-installing the latest this is now working! 🥳
