### 2025-06-24 03b5e26
update workflows to v4 of the artifact actions
