+++ .github/workflows/test_coverage.yml
@@ -20,7 +20,7 @@ jobs:
           r-version: '4.1.0'
       - uses: r-lib/actions/setup-pandoc@v2-branch
       - uses: r-lib/actions/setup-tinytex@v2-branch
-      - uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v4
 
       - name: Cache R packages
         uses: actions/cache@v3
