### 2025-06-29 #7697 @Nikoleta-v3
Thank you for your patience @evaaepelde, I’ve opened a pull request to tweak one of the references: https://github.com/bc3LC/medusa/pull/1. After this, we should be good to go!
