### 2025-06-30 #7697 @Nikoleta-v3
editorialbot create post-review checklist
