### 2025-06-30 #7697 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
