+++ paper/paper.bib
@@ -17,7 +17,7 @@
 }
 
 @article{cho2013,
-	title = {Toward a {Field} of {Intersectionality} {Studies}: {Theory}, {Applications}, and {Praxis}},
+	title = {Toward a field of intersectionality studies: {Theory}, applications, and praxis},
 	volume = {38},
 	issn = {0097-9740},
 	shorttitle = {Toward a {Field} of {Intersectionality} {Studies}},
@@ -36,7 +36,7 @@
 }
 
 @incollection{crenshaw1994,
-	title = {Mapping the {Margins}: {Intersectionality}, {Identity} {Politics}, and {Violence} {Against} {Women} of {Color}},
+	title = {Mapping the margins: {Intersectionality}, identity politics, and violence against women of color},
 	isbn = {978-0-203-06090-2},
 	shorttitle = {Mapping the {Margins}},
 	abstract = {Mapping the Margins: Intersectionality, Identity Politics, and Violence Against Women of Color - 1},
@@ -161,7 +161,7 @@
 }
 
 @techreport{eurostat2003,
-	title = {{HOUSEHOLD} {BUDGET} {SURVEYS} {IN} {THE} {EU} — {Methodology} and recommendations for harmonisation – 2003},
+	title = {{Household} budget surveys in the {EU} — {Methodology} and recommendations for harmonisation – 2003},
 	language = {en},
 	author = {Eurostat},
 	year = {2003},
