### 2025-07-01 #7697 @evaaepelde
Hi @samhforbes ,

Yes, I completely agree — that DOI suggestion seems incorrect, so happy to skip it.

Thanks also for the other notes! I’ve made the changes you mentioned: updated the article titles to sentence case and corrected the capitalisation for Eurostat (2003).

Let me know if anything else comes up! 😊
